+++ linux_termios.h
@@ -20,12 +20,14 @@
 #ifndef LIBSERIALPORT_LINUX_TERMIOS_H
 #define LIBSERIALPORT_LINUX_TERMIOS_H
 
-SP_PRIV int get_termios_get_ioctl(void);
-SP_PRIV int get_termios_set_ioctl(void);
-SP_PRIV int get_termios_size(void);
+#include <stdlib.h>
+
+SP_PRIV unsigned long get_termios_get_ioctl(void);
+SP_PRIV unsigned long get_termios_set_ioctl(void);
+SP_PRIV size_t get_termios_size(void);
 SP_PRIV int get_termios_speed(void *data);
 SP_PRIV void set_termios_speed(void *data, int speed);
-SP_PRIV int get_termiox_size(void);
+SP_PRIV size_t get_termiox_size(void);
 SP_PRIV int get_termiox_flow(void *data, int *rts, int *cts, int *dtr, int *dsr);
 SP_PRIV void set_termiox_flow(void *data, int rts, int cts, int dtr, int dsr);
